player: Ensure video retains consistent sizing
This ensures that cycling between full screen and non-full screen does
not cause any change of the positioning of the video nor its size.

By removing the video element CSS in main.scss, we can additionally
have a fully responsive video element that will be a nominal 600px for
the majority of our video content.  This can be explored later.

Closes: AP-339